requirements.txt
google-genai>=0.5.0
python-dotenv>=1.0
